Revert "Add zopfli and pigz repositories"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd20770d-ff00-493b-8e42-1a83ddc2aa5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I11e75637f4dfb8f25ec65b950440a79c1e081b9e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com